Refactor all components to be in separate directories
