+++ frontend/src/app/app-routing.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { WardrobeListComponent } from './wardrobe-list/wardrobe-list.component';
-import { GarmentDetailsComponent } from './garment-details/garment-details.component';
-import { GarmentFormComponent } from './garment-form/garment-form.component';
-import { UsageListComponent } from './usage-list/usage-list.component';
-import { UsageDetailsComponent } from './usage-details/usage-details.component';
-import { UsageFormComponent } from './usage-form/usage-form.component';
-import { ListingListComponent } from './listing-list/listing-list.component';
-import { ListingFormComponent } from './listing-form/listing-form.component';
-import { ListingDetailsComponent } from './listing-details/listing-details.component';
+import { WardrobeListComponent } from './components/garment/wardrobe-list/wardrobe-list.component';
+import { GarmentDetailsComponent } from './components/garment/garment-details/garment-details.component';
+import { GarmentFormComponent } from './components/garment/garment-form/garment-form.component';
+import { UsageListComponent } from './components/usage/usage-list/usage-list.component';
+import { UsageDetailsComponent } from './components/usage/usage-details/usage-details.component';
+import { UsageFormComponent } from './components/usage/usage-form/usage-form.component';
+import { ListingListComponent } from './components/listing/listing-list/listing-list.component';
+import { ListingFormComponent } from './components/listing/listing-form/listing-form.component';
+import { ListingDetailsComponent } from './components/listing/listing-details/listing-details.component';
 
 const routes: Routes = [
     // { path: '', redirectTo: '/wardrobe', pathMatch: 'full' },
