+++ frontend/src/app/components/garment/garment-details/garment-details.component.spec.ts
