+++ frontend/src/app/components/garment/garment-details/garment-details.component.ts
