+++ frontend/src/app/components/garment/garment-form/garment-form.component.ts
