+++ frontend/src/app/components/garment/wardrobe-list/wardrobe-list.component.spec.ts
