+++ frontend/src/app/components/listing/listing-details/listing-details.component.ts
