+++ frontend/src/app/components/listing/listing-list/listing-list.component.spec.ts
