+++ frontend/src/app/components/usage/usage-details/usage-details.component.spec.ts
