+++ frontend/src/app/components/usage/usage-details/usage-details.component.ts
